create cluster_entities.py and database_entities.py, move class definitions around, update test imports
